+++ syntax/tests/parsing/errors/typeDef/polyvariant.js
@@ -1,7 +1,5 @@
 type t = [< ]
 
-type t = [ s ]
-
 type z = [< | #A | #B > ]
 
 type rec t = []